+++ subtest-impl/tests/expand_impl/unused_mut_in_subtest.rs
@@ -1,0 +1,13 @@
+#[subtest]
+#[test]
+fn value_can_be_sent() {
+    let mut value = 1;
+
+    #[subtest]
+    fn another_value_can_be_sent() {
+        assert_eq!(value, 1);
+    }
+
+    value += 1;
+    assert_eq!(value, 2);
+}