+++ subtest/tests/integration_tests/expand/readme_inheritance_example.expanded.rs
@@ -254,6 +254,7 @@ mod value_can_be_sent_and_received_subtests {
     #[should_panic(expected = "called `Result::unwrap()` on an `Err` value: Empty")]
     fn value_cannot_be_received_a_second_time() {
         #[allow(unused_variables)]
+        #[allow(unused_mut)]
         let (sender, mut receiver) = tokio::sync::mpsc::channel(5);
         sender.try_send("Hello!").unwrap();
         receiver.try_recv().unwrap();
@@ -287,6 +288,7 @@ mod value_can_be_sent_and_received_subtests {
     #[ignore]
     fn value_can_be_sent_a_second_time() {
         #[allow(unused_variables)]
+        #[allow(unused_mut)]
         let (sender, mut receiver) = tokio::sync::mpsc::channel(5);
         sender.try_send("Hello!").unwrap();
         receiver.try_recv().unwrap();